# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoraPay API",
    "version": "0.1.0",
    "description": "Cross-border remittance platform. Russia/CIS → Nigeria & Ghana.\n\n**No live funds.** `LIVE_FUNDS_ENABLED` defaults to false and every provider\nruns against a simulator until partner agreements and written legal opinions\nare in place.\n\nTwo conventions that differ from most payment APIs, and both are deliberate:\n\n1. **Money is never a JSON number.** Amounts carry `minorUnits` as a string,\n   which is the authoritative field. A JSON number is a double, and a double\n   is not a currency.\n2. **An acknowledgement is not a settlement.** A provider that accepts a\n   submission has moved no money. Only a status poll or the T+1 statement\n   produces an outcome, and only an outcome posts to the ledger.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.staging.morapay.example",
      "description": "Staging"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Back-office tokens are signed with a different key from customer tokens. A customer token fails signature verification here, not merely authorisation."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    },
...
  },
  "paths": {
...
    "/recipients/name-enquiry": {
      "post": {
        "tags": [
          "Recipients"
        ],
        "summary": "Resolve the name the institution holds",
        "description": "Called before the sender commits. Showing the resolved name back to the sender prevents the most common support case in this corridor: money sent to a mistyped account number, unrecoverable once settled.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "details": {
                    "oneOf": [
                      {
                        "type": "object",
                        "properties": {
                          "method": {
                            "type": "string",
                            "enum": [
                              "BANK_ACCOUNT"
                            ]
                          },
                          "country": {
                            "type": "string",
                            "enum": [
                              "NG"
                            ]
                          },
                          "accountNumber": {
                            "type": "string",
                            "pattern": "^\\d{10}$"
                          },
                          "bankCode": {
                            "type": "string",
                            "pattern": "^\\d{3}$"
                          },
                          "declaredName": {
                            "type": "string",
                            "minLength": 2,
                            "maxLength": 120
                          }
                        },
                        "required": [
                          "method",
                          "country",
                          "accountNumber",
                          "bankCode",
                          "declaredName"
                        ]
                      },
                      {
                        "type": "object",
                        "properties": {
                          "method": {
                            "type": "string",
                            "enum": [
                              "MOBILE_MONEY"
                            ]
                          },
                          "country": {
                            "type": "string",
                            "enum": [
                              "GH"
                            ]
                          },
                          "msisdn": {
                            "type": "string",
                            "pattern": "^233\\d{9}$"
                          },
                          "network": {
                            "type": "string",
                            "enum": [
                              "MTN",
                              "TELECEL",
                              "AIRTELTIGO"
                            ]
                          },
                          "declaredName": {
                            "type": "string",
                            "minLength": 2,
                            "maxLength": 120
                          }
                        },
                        "required": [
                          "method",
                          "country",
                          "msisdn",
                          "network",
                          "declaredName"
                        ]
                      }
                    ]
                  }
                },
                "required": [
                  "details"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolved or not",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "RESOLVED"
                          ]
                        },
                        "resolvedName": {
                          "type": "string"
                        },
                        "institution": {
                          "type": "string"
                        },
                        "matchesDeclaredName": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "status",
                        "resolvedName",
                        "institution",
                        "matchesDeclaredName"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "NOT_FOUND"
                          ]
                        },
                        "reason": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "status",
                        "reason"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "UNSUPPORTED"
                          ]
                        },
                        "reason": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "status",
                        "reason"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — unverified email, insufficient role, or a limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "A domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/recipients": {
      "post": {
        "tags": [
          "Recipients"
        ],
        "summary": "Save a recipient",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "details": {
                    "oneOf": [
                      {
                        "type": "object",
                        "properties": {
                          "method": {
                            "type": "string",
                            "enum": [
                              "BANK_ACCOUNT"
                            ]
                          },
                          "country": {
                            "type": "string",
                            "enum": [
                              "NG"
                            ]
                          },
                          "accountNumber": {
                            "type": "string",
                            "pattern": "^\\d{10}$"
                          },
                          "bankCode": {
                            "type": "string",
                            "pattern": "^\\d{3}$"
                          },
                          "declaredName": {
                            "type": "string",
                            "minLength": 2,
                            "maxLength": 120
                          }
                        },
                        "required": [
                          "method",
                          "country",
                          "accountNumber",
                          "bankCode",
                          "declaredName"
                        ]
                      },
                      {
                        "type": "object",
                        "properties": {
                          "method": {
                            "type": "string",
                            "enum": [
                              "MOBILE_MONEY"
                            ]
                          },
                          "country": {
                            "type": "string",
                            "enum": [
                              "GH"
                            ]
                          },
                          "msisdn": {
                            "type": "string",
                            "pattern": "^233\\d{9}$"
                          },
                          "network": {
                            "type": "string",
                            "enum": [
                              "MTN",
                              "TELECEL",
                              "AIRTELTIGO"
                            ]
                          },
                          "declaredName": {
                            "type": "string",
                            "minLength": 2,
                            "maxLength": 120
                          }
                        },
                        "required": [
                          "method",
                          "country",
                          "msisdn",
                          "network",
                          "declaredName"
                        ]
                      }
                    ]
                  },
                  "nickname": {
                    "type": "string",
                    "maxLength": 60
                  }
                },
                "required": [
                  "details"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string",
                      "enum": [
                        "BANK_ACCOUNT",
                        "MOBILE_MONEY"
                      ]
                    },
                    "country": {
                      "type": "string",
                      "enum": [
                        "RU",
                        "BY",
                        "NG",
                        "GH"
                      ]
                    },
                    "maskedAccount": {
                      "type": "string"
                    },
                    "resolvedName": {
                      "type": "string",
                      "nullable": true
                    },
                    "nickname": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "method",
                    "country",
                    "maskedAccount",
                    "resolvedName",
                    "nickname",
                    "createdAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — unverified email, insufficient role, or a limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "A domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Recipients"
        ],
        "summary": "List saved recipients",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recipients": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "method": {
                            "type": "string",
                            "enum": [
                              "BANK_ACCOUNT",
                              "MOBILE_MONEY"
                            ]
                          },
                          "country": {
                            "type": "string",
                            "enum": [
                              "RU",
                              "BY",
                              "NG",
                              "GH"
                            ]
                          },
                          "maskedAccount": {
                            "type": "string"
                          },
                          "resolvedName": {
                            "type": "string",
                            "nullable": true
                          },
                          "nickname": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "method",
                          "country",
                          "maskedAccount",
                          "resolvedName",
                          "nickname",
                          "createdAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "recipients"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — unverified email, insufficient role, or a limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "A domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}